requirements.txt
aiogram>=3.0.0
google-genai
asyncpg
python-dotenv
pillow
